Move PROTOCOL and FASTQ_SCREEN constants to new qc/constants.py module.
